requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
decli==0.6.3
deprecation==2.1.0
diff-match-patch==20241021
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-import-export==4.3.9
django-jazzmin==3.0.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
filelock==3.18.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
Markdown==3.8.2
MarkupSafe==3.0.2
nodeenv==1.9.1
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
postgrest==1.1.1
pre_commit==4.2.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
questionary==2.1.0
realtime==2.6.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
tablib==3.8.0
termcolor==3.1.0
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.32.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.9.0
zipp==3.23.0
google-api-python-client==2.147.0
yt-dlp==2024.8.6
